perf(gpu): halve quant_levels buffer via int16_t, and stop dispatching a dead entropy shader
quant_levels held quantized H.264 transform coefficients as `int`
(4 bytes) even though every real value fits in int16_t: dequantized
16-bit residuals scaled by a 4-bit qp_scale factor never approach the
2^31 range, and the encoder already clamped writes into it. At 1440p
this buffer is 22 MB per context; halving it cuts that GPU-CPU staging
copy and the CPU-side CAVLC/CABAC read traffic proportionally.

Changed shaders (quantize.comp, intra_wavefront.comp, reconstruct.comp,
deblock_filter.comp) to declare the SSBO as int16_t via
GL_EXT_shader_16bit_storage + GL_EXT_shader_explicit_arithmetic_types_int16
(device support for storageBuffer16BitAccess/shaderInt16 already
verified present on this board). Threaded int16_t through
gpu_compute.c's size computation, encoder_h264.c's quant_levels_shadow
and every accessor, and cavlc.c/.h's block-encode signatures, leaving
coeff_buffer/residual_buffer/dc_coeff_buffer untouched since they carry
different, wider-range data.

The compiler's type checking caught two real missed call sites that a
grep-based pass would have silently gotten wrong: the CAVLC-path luma
DC block (dc_out, a plain int[16] from the Hadamard stage) and an
all-zero fallback block (zero16). Both now narrow-copy into a local
int16_t[16] rather than changing the shared Hadamard-side types.

Also found and fixed two real bugs surfaced while doing this:

- BC250_NZ_AUDIT's buffer-size guard still checked `>= ... * sizeof(int)`
  against a buffer that is now half that size, which would have made
  the audit silently never run instead of verifying anything. Fixed to
  sizeof(int16_t).

- gpu_compute.c was unconditionally dispatching the entropy_encode.comp
  compute shader every frame, gated only on "did the pipeline load", not
  "does anything still consume its output" - nothing has read from
  entropy_buffer since CAVLC/CABAC moved to real host-side entropy
  coding. This was ~0.34ms/frame of wasted GPU time on every frame, and
  after halving quant_levels_buffer it would have read out of bounds
  against the resized buffer. Removed the dispatch and its dependent
  vkCmdCopyBuffer; pipeline/buffer objects are left allocated but never
  invoked.

Verified via tools/lab gate against the pre-change commit:
- all 4 unit test binaries pass (test_bitstream, test_cavlc, test_encode,
  test_va_api)
- BC250_NZ_AUDIT: 0 mismatches/200 frames on both testsrc and testsrc2
- quality: PSNR ~58.2-58.4dB, SSIM 0.999 on testsrc2
- byte-identical vs baseline on testsrc at gop=120, gop=10, and gop=1
  (all-intra) - this change is a pure memory-layout optimization with
  zero effect on the encoded bitstream, confirmed on the I-frame-only
  path specifically per this project's practice of checking that path
  separately

+++ approach1-compute-encoder/src/cavlc.c
@@ -605,7 +605,7 @@ static void cavlc_write_run_befores(bitstream_t *bs, const int *runs, int total_
     }
 }
 
-int cavlc_write_4x4_block(bitstream_t *bs, const int *coeffs, int nC) {
+int cavlc_write_4x4_block(bitstream_t *bs, const int16_t *coeffs, int nC) {
     if (!bs || !coeffs) return 0;
 
     int scanned[16];
@@ -688,7 +688,7 @@ int cavlc_write_4x4_block(bitstream_t *bs, const int *coeffs, int nC) {
  * (see cavlc_write_total_zeros's comment) with the max_coeff=15 "omit when
  * full" threshold; coeff_token and run_before are unchanged/shared.
  */
-int cavlc_write_4x4_ac_block(bitstream_t *bs, const int *coeffs, int nC) {
+int cavlc_write_4x4_ac_block(bitstream_t *bs, const int16_t *coeffs, int nC) {
     if (!bs || !coeffs) return 0;
 
     int scanned[15];
